+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Changelog
 
+# Unreleased
+
+- Upgraded node-llama-cpp version
+- Improved error and log messages
+- Changed startup to remember the selected model without loading it until the
+  first message
+- Added a setting to load the selected model automatically on startup
+
 # 0.1.0 - Initial release